+++ static-webapps/slate-portfolio-manager/src/router.js
@@ -1,7 +1,6 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 
-import { useAuth } from '@/store/auth';
 import AdvancedPortfolioManager from './views/AdvancedPortfolioManager.vue';
 
 const routes = [
