+++ static-webapps/slate-portfolio-manager/src/store/client.js
@@ -1,8 +1,16 @@
 import axios from 'axios';
 
+export const redirectToLogin = () => {
+  const next = encodeURIComponent(window.location.href)
+  window.location = 'http://localhost:2190/login?return=' + next
+}
+
 export const getClient = (options = {}) => {
-  const handleError = (e) => {
-    throw e;
+  const handleError = (error) => {
+    if (error.response?.status === 401) {
+      redirectToLogin()
+    }
+    throw error;
   };
   const client = axios.create({
     baseURL: options.baseURL,
