+++ static-webapps/slate-portfolio-manager/src/store/useContentArea.js
@@ -1,0 +1,6 @@
+import defineRestStore from './defineRestStore';
+
+export default defineRestStore({
+  id: 'contentArea',
+  baseURL: '/cbl/content-areas',
+});
